refactor(content): WKC layout — Phase 7 (sync subsystem)
Phase 7 of the WKC migration. Reshapes the synchronization subsystem
(the last big chunk of legacy src/service/) into the WKC layout.

New / moved components:

- adapters/synchronization-state/ — moved from
  ports/synchronizationState.ts. Holds the bootstrapping->syncing state
  enum plus listeners; rules-free state, so it stays an adapter.
- logic/challenge-supervisor/ — moved from
  service/synchronization/ChallengeSupervisor.ts. Class shape preserved.
- logic/retry-failed-deployments/ — moved from
  service/synchronization/retryFailedDeployments.ts.
- logic/batch-deployer/ — moved from
  service/synchronization/batchDeployer.ts (237 LoC). Imports
  deployEntityFromRemoteServer through the sync-orchestrator index so
  jest.spyOn on the module re-export still intercepts the call.
- logic/sync-orchestrator/ — combines logic/synchronization.ts
  (startSynchronization) and service/synchronization/deployRemoteEntity.ts
  (downloadFullEntity, deployDownloadedEntity, deployEntityFromRemoteServer).
  The deploy-remote-entity helpers are re-exported from index.ts so other
  components can mock them via the canonical module path.
- logic/peer-cluster/ — renamed from logic/cluster/. Same component
  factory and surface; the directory rename is the WKC convention
  alignment (cluster is too generic; peer-cluster names the actual
  responsibility — discovering and tracking peer catalysts via DAO).

Wiring:

- All importers swept via sed:
  ports/synchronizationState -> adapters/synchronization-state
  service/synchronization/* -> logic/* per the table above
  logic/synchronization -> logic/sync-orchestrator
  logic/cluster -> logic/peer-cluster
- src/service/synchronization/ deleted entirely.
- src/logic/cluster/ deleted (replaced by peer-cluster).
- src/service/ now contains only garbage-collection/ (Phase 8) and
  time/ (a tiny helper).

Behavior preserved:

- Same start/stop semantics on every component.
- Same DAO poll interval and bootstrap flow.
- Same batch-deployer concurrency and priority logic.
- Same retry schedule and exponential-backoff timing.

tsc clean, unit suite green (213 tests, no diff).
