+++ content/src/adapters/synchronization-state/index.ts
@@ -1,0 +1,3 @@
+export { createSynchronizationState } from './component'
+export { State } from './types'
+export type { SynchronizationState } from './types'
