+++ content/src/adapters/synchronization-state/types.ts
@@ -1,0 +1,9 @@
+export enum State {
+  BOOTSTRAPPING = 'Bootstrapping',
+  SYNCING = 'Syncing'
+}
+
+export interface SynchronizationState {
+  getState: () => State
+  toSyncing: () => void
+}
