+++ content/src/components.ts
@@ -36,14 +36,14 @@ import { createProcessedSnapshotStorage } from './ports/processedSnapshotStorage
 import { createSequentialTaskExecutor } from './adapters/sequential-task-executor'
 import { createSnapshotGenerator } from './ports/snapshotGenerator'
 import { createSnapshotStorage } from './ports/snapshotStorage'
-import { createSynchronizationState } from './ports/synchronizationState'
+import { createSynchronizationState } from './adapters/synchronization-state'
 import { createSystemProperties } from './adapters/system-properties'
 import { GarbageCollectionManager } from './service/garbage-collection/GarbageCollectionManager'
 import { PointerManager } from './logic/pointer-manager'
-import { ChallengeSupervisor } from './service/synchronization/ChallengeSupervisor'
-import { createContentCluster } from './logic/cluster'
-import { createBatchDeployerComponent } from './service/synchronization/batchDeployer'
-import { createRetryFailedDeployments } from './service/synchronization/retryFailedDeployments'
+import { ChallengeSupervisor } from './logic/challenge-supervisor'
+import { createContentCluster } from './logic/peer-cluster'
+import { createBatchDeployerComponent } from './logic/batch-deployer'
+import { createRetryFailedDeployments } from './logic/retry-failed-deployments'
 import { createContentValidator } from './adapters/content-validator'
 import { createAuthenticator } from './logic/authenticator'
 import { createServerValidator } from './logic/server-validator'
