+++ content/src/logic/batch-deployer/index.ts
@@ -1,0 +1,1 @@
+export { createBatchDeployerComponent, priorityBasedOnEntityType } from './component'
