+++ content/src/logic/challenge-supervisor/index.ts
@@ -1,0 +1,2 @@
+export { ChallengeSupervisor } from './component'
+export type { IChallengeSupervisor } from './types'
