+++ content/src/logic/challenge-supervisor/types.ts
@@ -1,0 +1,4 @@
+export type IChallengeSupervisor = {
+  getChallengeText(): string
+  isChallengeOk(text: string): boolean
+}
