--- content/src/logic/cluster/index.ts
@@ -1,2 +1,0 @@
-export { createContentCluster } from './component'
-export { IContentClusterComponent } from './types'
