+++ content/src/logic/peer-cluster/index.ts
@@ -1,0 +1,2 @@
+export { createContentCluster } from './component'
+export type { IContentClusterComponent } from './types'
