+++ content/src/logic/retry-failed-deployments/index.ts
@@ -1,0 +1,2 @@
+export { createRetryFailedDeployments } from './component'
+export type { IRetryFailedDeploymentsComponent } from './types'
