+++ content/src/logic/retry-failed-deployments/types.ts
@@ -1,0 +1,5 @@
+import { IBaseComponent } from '@well-known-components/interfaces'
+
+export type IRetryFailedDeploymentsComponent = IBaseComponent & {
+  schedule: () => Promise<void>
+}
