+++ content/src/logic/sync-orchestrator/index.ts
@@ -1,0 +1,2 @@
+export { startSynchronization } from './component'
+export { deployEntityFromRemoteServer, deployDownloadedEntity } from './deploy-remote-entity'
