+++ content/src/service.ts
@@ -1,7 +1,7 @@
 import { IHttpServerComponent, Lifecycle } from '@well-known-components/interfaces'
 import { setupRouter } from './controller/routes'
 import { EnvironmentConfig } from './Environment'
-import { startSynchronization } from './logic/synchronization'
+import { startSynchronization } from './logic/sync-orchestrator'
 import { migrateContentFolderStructure } from './migrations/ContentFolderMigrationManager'
 import { AppComponents, GlobalContext } from './types'
 import path from 'path'
