+++ content/src/types.ts
@@ -37,13 +37,13 @@ import { IFailedDeploymentsComponent } from './ports/failedDeployments'
 import { IDatabaseComponent } from './ports/postgres'
 import { ISequentialTaskExecutorComponent } from './adapters/sequential-task-executor'
 import { SnapshotGenerator } from './ports/snapshotGenerator'
-import { SynchronizationState } from './ports/synchronizationState'
+import { SynchronizationState } from './adapters/synchronization-state'
 import { SystemProperties } from './adapters/system-properties'
 import { GarbageCollectionManager } from './service/garbage-collection/GarbageCollectionManager'
 import { PointerManager } from './logic/pointer-manager'
-import { IChallengeSupervisor } from './service/synchronization/ChallengeSupervisor'
-import { IContentClusterComponent } from './logic/cluster'
-import { IRetryFailedDeploymentsComponent } from './service/synchronization/retryFailedDeployments'
+import { IChallengeSupervisor } from './logic/challenge-supervisor'
+import { IContentClusterComponent } from './logic/peer-cluster'
+import { IRetryFailedDeploymentsComponent } from './logic/retry-failed-deployments'
 import { ProcessedSnapshotsStorageComponent } from './ports/processedSnapshotStorage'
 import ms from 'ms'
 import { IDeploymentsComponent } from './logic/deployments'
