+++ content/test/integration/TestProgram.ts
@@ -6,7 +6,7 @@ import fetch from 'node-fetch'
 import { EnvironmentConfig } from '../../src/Environment'
 import { AuditInfo, Deployment, DeploymentOptions, isInvalidDeployment } from '../../src/deployment-types'
 import { getDeployments } from '../../src/logic/deployments'
-import * as synchronization from '../../src/logic/synchronization'
+import * as synchronization from '../../src/logic/sync-orchestrator'
 import { FailedDeployment } from '../../src/ports/failedDeployments'
 import { main } from '../../src/service'
 import { AppComponents } from '../../src/types'
